Moved most of the logic to ship and implemented notification
Mostly this was to help with networking. If we are going to be running the game state on a server, we need it seperate from these game objects which appear to do too much. If that means reimplementing my own version of chingu objects then so be it. Better I learn it this way than get too attached to a framework.